First Attempt
annscript3 has k fold cross validation and grid search to try to find
the best network topology.
Best one so far was 40 neurons per net, with 4 layers.
Note the error was based off of binary targets (i.e. whichever emotion
was the dominant one was marked with 1, then remainder was marked with
0)